Fix build error after embedded framework changed
Say you have an app that uses an embedded framework:

    # Rakefile
    Motion::Project::App.setup do |app|
      app.name = 'ABC'
      app.embedded_frameworks << "vendor/XYZ.framework"
    end

When you run `rake` it copies the framework folder into the app bundle:

    build/.../ABC.app/Contents/Frameworks/XYZ.framework/

But when the framework changes -- for example when its recompiled, or
updated to a newer version -- running `rake` again copies the newer
version into a subfolder:

    build/.../ABC.app/Contents/Frameworks/XYZ.framework/XYZ.framework

The result is that the app still links against the older version.

Now when you run `rake` a third time, you get an error message along the
lines of:

    Errno::EEXIST: File exists @ sys_fail2 - (Versions/Current/Headers,
    ./build/.../ABC.app/Contents/Frameworks/XYZ.framework/XYZ.framework/Headers)

This commit fixes this problem by removing the old framework if a newer
version is available.