Frameworks require all embedded code to be signed.
On OS X, if we're signing a vendored framework, then we include the `--deep` flag on the codesign to ensure that any nested code content such as helpers, frameworks, and plug-ins are also signed (requirement since Mavericks).  The same identity to sign the framework is used to the sign the nested code.